% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{list.dirs.depth.n}
\alias{list.dirs.depth.n}
\title{list directories to depth of n}
\usage{
list.dirs.depth.n(p, n)
}
\arguments{
\item{p}{path}

\item{n}{number of directories to traverse down (as in find's maxdepth)}
}
\value{
vector of directory names
}
\description{
list directories to depth of n
}
\details{
from https://stackoverflow.com/questions/48297440/list-files-recursive-up-to-a-certain-level-in-r
}
\examples{
list.dirs.depth.n(".", n = 3)
}
